+++ Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/model/ChanceLocation.java
@@ -22,14 +22,14 @@ package org.l2j.gameserver.model;
  * @author UnAfraid
  */
 public class ChanceLocation extends Location {
-    private final double _chance;
+    private final float _chance;
 
-    public ChanceLocation(int x, int y, int z, int heading, double chance) {
+    public ChanceLocation(int x, int y, int z, int heading, float chance) {
         super(x, y, z, heading);
         _chance = chance;
     }
 
-    public double getChance() {
+    public float getChance() {
         return _chance;
     }
 }
